requirements.txt
tensorflow
datasets
numpy
scikit-learn
pillow
pyyaml
dvc
